CheckSjisChar 関数の第2引数の型を int から size_t に変更

+++ sakura_core/charset/codechecker.h
@@ -408,7 +408,7 @@ inline int GuessEucjpCharsz( const char uc_ ){
 	文字長検査
 */
 /* --- ローカル文字コードチェック */
-int CheckSjisChar( const char*, const int, ECharSet* );
+int CheckSjisChar( const char*, const size_t, ECharSet* );
 int CheckEucjpChar( const char*, const int, ECharSet* );
 int DetectJisEscseq( const char*, const size_t, EMyJisEscseq* ); // JIS エスケープシーケンス検出器
 int _CheckJisAnyPart( const char*, const int, const char **ppNextChar, EMyJisEscseq *peNextEsc, int *pnErrorCount, const int nType );
